Simplify - we don't need to build matplotlib-cpp, it's header-only
The matplotlib-cpp subproject is:
- The matplotlibcpp.h header file
- A bunch of examples and CMakeLists to build the examples
  against the header.

It's nice to build the examples and I learned a lot by making it work
but in the end, I just need matplotlibcpp.h.  Simplify the Dockerfile
to avoid the "Builder" stage and just copy the headers into the docker
image that is set up with xeus-cling.